refactor: move all hooks to dedicated folder
