+++ packages/outstatic/src/components/MetadataBuilder/index.tsx
@@ -19,7 +19,7 @@ import { useApollo } from '../../utils/apollo'
 import matter from 'gray-matter'
 import { chunk } from '../../utils/chunk'
 import { hashFromUrl } from '../../utils/hashFromUrl'
-import useOid from '../../utils/useOid'
+import useOid from '../../utils/hooks/useOid'
 import { createCommit } from '../../utils/createCommit'
 import MurmurHash3 from 'imurmurhash'
 import { MetadataSchema, OutstaticSchema } from '../../utils/metadata/types'
