+++ packages/outstatic/src/utils/hooks/useNavigationLock.tsx
